Refactor paginated search widgets
- Implement a base class
- Override specific members
