+++ lib/widget/part_list.dart
@@ -1,14 +1,15 @@
-import 'package:flutter/material.dart';
-import 'package:infinite_scroll_pagination/infinite_scroll_pagination.dart';
-import 'package:inventree/inventree/part.dart';
-import 'package:inventree/inventree/sentry.dart';
-import 'package:inventree/widget/paginator.dart';
-import 'package:inventree/widget/part_detail.dart';
-import 'package:inventree/widget/refreshable_state.dart';
+import "package:flutter/material.dart";
+import "package:infinite_scroll_pagination/infinite_scroll_pagination.dart";
+import 'package:inventree/inventree/model.dart';
+import "package:inventree/inventree/part.dart";
+import "package:inventree/inventree/sentry.dart";
+import "package:inventree/widget/paginator.dart";
+import "package:inventree/widget/part_detail.dart";
+import "package:inventree/widget/refreshable_state.dart";
 
-import '../api.dart';
-import '../app_settings.dart';
-import '../l10.dart';
+import "package:inventree/api.dart";
+import "package:inventree/app_settings.dart";
+import "package:inventree/l10.dart";
 
 
 class PartList extends StatefulWidget {
@@ -52,86 +53,21 @@ class PaginatedPartList extends StatefulWidget {
 }
 
 
-class _PaginatedPartListState extends State<PaginatedPartList> {
+class _PaginatedPartListState extends PaginatedSearchState<PaginatedPartList> {
 
-  _PaginatedPartListState(this.filters, this.onTotalChanged);
-
-  static const _pageSize = 25;
-
-  String _searchTerm = "";
+  _PaginatedPartListState(Map<String, String> filters, this.onTotalChanged) : super(filters);
 
   Function(int)? onTotalChanged;
 
-  final Map<String, String> filters;
-
-  final PagingController<int, InvenTreePart> _pagingController = PagingController(firstPageKey: 0);
-
   @override
-  void initState() {
-    _pagingController.addPageRequestListener((pageKey) {
-      _fetchPage(pageKey);
-    });
+  Future<InvenTreePageResponse?> requestPage(int limit, int offset, Map<String, String> params) async {
+    final bool cascade = await InvenTreeSettingsManager().getBool("partSubcategory", true);
 
-    super.initState();
-  }
+    params["cascade"] = "${cascade}";
 
-  @override
-  void dispose() {
-    _pagingController.dispose();
-    super.dispose();
-  }
+    final page = await InvenTreePart().listPaginated(limit, offset, filters: params);
 
-  int resultCount = 0;
-
-  Future<void> _fetchPage(int pageKey) async {
-    try {
-
-      Map<String, String> params = filters;
-
-      params["search"] = _searchTerm;
-
-      final bool cascade = await InvenTreeSettingsManager().getBool("partSubcategory", true);
-
-      params["cascade"] = "${cascade}";
-
-      final page = await InvenTreePart().listPaginated(_pageSize, pageKey, filters: params);
-      int pageLength = page?.length ?? 0;
-      int pageCount = page?.count ?? 0;
-
-      final isLastPage = pageLength < _pageSize;
-
-      // Construct a list of part objects
-      List<InvenTreePart> parts = [];
-
-      if (page != null) {
-        for (var result in page.results) {
-          if (result is InvenTreePart) {
-            parts.add(result);
-          }
-        }
-      }
-
-      if (isLastPage) {
-        _pagingController.appendLastPage(parts);
-      } else {
-        final int nextPageKey = pageKey + pageLength;
-        _pagingController.appendPage(parts, nextPageKey);
-      }
-
-      if (onTotalChanged != null) {
-        onTotalChanged!(pageCount);
-      }
-
-      setState(() {
-        resultCount = pageCount;
-      });
-
-    } catch (error, stackTrace) {
-      print("Error! - ${error.toString()}");
-      _pagingController.error = error;
-
-      sentryReportError(error, stackTrace);
-    }
+    return page;
   }
 
   void _openPart(BuildContext context, int pk) {
@@ -144,7 +80,11 @@ class _PaginatedPartListState extends State<PaginatedPartList> {
     });
   }
 
-  Widget _buildPart(BuildContext context, InvenTreePart part) {
+  @override
+  Widget buildItem(BuildContext context, InvenTreeModel model) {
+
+    InvenTreePart part = model as InvenTreePart;
+
     return ListTile(
       title: Text(part.fullname),
       subtitle: Text("${part.description}"),
@@ -157,45 +97,6 @@ class _PaginatedPartListState extends State<PaginatedPartList> {
       onTap: () {
         _openPart(context, part.pk);
       },
-    );
-  }
-
-  final TextEditingController searchController = TextEditingController();
-
-  void updateSearchTerm() {
-
-    _searchTerm = searchController.text;
-    _pagingController.refresh();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      mainAxisAlignment: MainAxisAlignment.start,
-      children: [
-        PaginatedSearchWidget(searchController, updateSearchTerm, resultCount),
-        Expanded(
-            child: CustomScrollView(
-              shrinkWrap: true,
-              physics: ClampingScrollPhysics(),
-              scrollDirection: Axis.vertical,
-              slivers: [
-                PagedSliverList.separated(
-                  pagingController: _pagingController,
-                  builderDelegate: PagedChildBuilderDelegate<InvenTreePart>(
-                    itemBuilder: (context, item, index) {
-                      return _buildPart(context, item);
-                    },
-                    noItemsFoundIndicatorBuilder: (context) {
-                      return NoResultsWidget(L10().partNoResults);
-                    },
-                  ),
-                  separatorBuilder: (context, index) => const Divider(height: 1),
-                )
-              ],
-            )
-        )
-      ],
     );
   }
 }
